+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+* Fixed int coercion in tryFromAny()
+* Fixed string comparison with ucfirst(name)
+
 ## v0.1.0 (2024-06-17)
 * Added fromAny and tryFromAny